Use local fusion server
Using a dedicated py3o fusion server is recommended for scalability.
In practice, we don't need much resources to generate the project report
for this project.

At the moment, the py3o fusion server suffers from a bug which does not
happen in the local fusion server. https://github.com/OCA/reporting-engine/issues/121

Pull requests have been proposed in the py3o.fusion lib and the
report_py3o project to be able to use the fusion server, but in the
meanwhile it is just better to use the local fusion server.

https://bitbucket.org/faide/py3o.fusion/pull-requests/1/add-templating-options-in-the-form/diff
https://github.com/OCA/reporting-engine/pull/122