[imp] better html formatting for project report BSMTS-103
Prior to this change the content of `conclusion` field
was rendered w/ visible html markup.
Now we strip markup out and we keep only line breaks
whereas a `<p />` was present.
We tried hard to get real html or to inject a complete odt markup
but is not workin as it gets always escaped.